Changed the test dataset size from 4 kiB to 256 MiB, for now.
This makes the test dataset not fit into any CPU caches, which is one interesting comparison point.  Although I think I want to parameterise the benchmark to gather data from a variety of points on the cache-memory curve, as there's some interesting variations in behaviour as data set size scales.