`Datas` now initialises itself fully in init, rather than deferring generation of the test data until first use.
This is mainly for simplicity.  No need for the complexity and (probably trivial) overhead of a lazy var when the data is definitely going to be needed imminently anyway.

Also, removed use of Array(unsafeUninitializedCapacity:…) for the outer array, as it appears its use was causing memory corruption!  I vaguely recall, now, that there's some warning in the docs about using that method with non-trivial data types.  So I think it's still safe for the inner arrays (of simple `Int`).