Now use the same colour for algorithm pairs that differ only by dictionary lookup vs linear search.
The order of algorithms - as sorted naturally - now matters, as such pairs need to be adjacent in the data given to Swift Charts, otherwise they'll be shown out of order or separated in the chart legend.

Also, increased the line width from two to three, and tweaked the dashed line style to look symmetrical (and visually clear) in the chart legend.