Added a chart comparing the algorithms across inputs.
Currently this conflates ASCII and non-ASCII inputs, which isn't the worst thing but does make the chart a little harder to interpret.  I should probably add the option to focus on just one ASCIIness mode.

It's currently taking the mean of duration per input byte (and should reflect any filtering performed on the X range, although I haven't checked that).  It might be better to present a more nuanced graph type for that - e.g. a box plot with the mean marked (whiskers don't make sense in this case since there's only a handful of input categories - not enough for a statistical breakdown of stddev etc).