Added the ability to drag the chart, providing both a file representation and direct data representations (in most image formats supported by NSImage).
The direct image data is provided lazily, which is ideal.  Unfortunately the Finder craps itself - it literally crashes - if you try to drop a pasteboard containing only direct image data.  The Finder requires the pasteboard contain a file URL.  So, the code pro-actively writes the TIFF version to a file.  Just for the fucking Finder. 😤