Adjusted the Y axis title spacing, from -10 sometimes to 10 universally, for the input charts.
This makes it uglier in the default, unlimited view (because of that "1,000s" axis label that's relatively wide), but better (no longer overlaps axis title with axis labels) in most cases.

It's shit that Swift Charts doesn't have a more intelligent layout engine, that understands that not all axis labels are the same size. 😕